+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # node-google-dfp-wrapper
 
+### 1.3.0
+- Adds `getCriteriaValues` for querying criteria values as full objects.
+- Adds `getCriteriaValueId` for querying criteria values and returning the first id.
+- Deprecates `getCriteriaValue` for querying criteria values and returning the first id.
+
 ### 1.2.1
 - Updated node-google-dfp to 0.2.1